Moving Pipeline to a new file
